FIX improve check on overwriting

+++ _updPub.sh
@@ -77,6 +77,10 @@ update_script() {
   local filename=${url##*/}
 
   get_file $url /tmp/${filename}.new
+  result=$?
+  if [[ $result -ne 0 ]]; then
+    return $result
+  fi
   cp /tmp/${filename}.new ${filename}
   rm /tmp/${filename}.new
 }
